+++ packages/serverless-fc-starter/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.74](https://github.com/midwayjs/midway-faas/compare/v0.2.73...v0.2.74) (2020-04-13)
+
+
+### Bug Fixes
+
+* fc starter apigw ([00da6a9](https://github.com/midwayjs/midway-faas/commit/00da6a94292b9217da94fde81c338666d44bde33))
+* support apigw ([abfc27a](https://github.com/midwayjs/midway-faas/commit/abfc27a93ae1b335bcb040679bbb5d39f71f5c9e))
+
+
+
+
+
+## [0.2.73](https://github.com/midwayjs/midway-faas/compare/v0.2.73-alpha.0...v0.2.73) (2020-04-11)
+
+**Note:** Version bump only for package @midwayjs/serverless-fc-starter
+
+
+
+
+
+## [0.2.73-alpha.0](https://github.com/midwayjs/midway-faas/compare/v0.2.72...v0.2.73-alpha.0) (2020-04-11)
+
+
+### Bug Fixes
+
+* fix buffer event ([2d43e75](https://github.com/midwayjs/midway-faas/commit/2d43e75b4c1c8561345a3ce00c23e55f76e34f0c))
+
+
+
+
+
+## [0.2.72](https://github.com/midwayjs/midway-faas/compare/v0.2.71...v0.2.72) (2020-04-11)
+
+
+### Bug Fixes
+
+* aliyun gateway accept number ([03a79b7](https://github.com/midwayjs/midway-faas/commit/03a79b74045dff1ed58522154025af3747549a13))
+
+
+
+
+
 ## [0.2.65](https://github.com/midwayjs/midway-faas/compare/v0.2.64...v0.2.65) (2020-04-05)
 
 **Note:** Version bump only for package @midwayjs/serverless-fc-starter